backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Google AI
google-generativeai==0.8.3

# ElevenLabs
elevenlabs

# Utilities
python-dotenv==1.0.1
pydantic==2.10.4
httpx==0.28.1

# CORS
python-multipart==0.0.20

